// src/app/page/master/map-kalender-to-hari-libur/map-kalender-to-hari-libur.interface.ts
import { EventEmitter } from '@angular/core';

export interface MapKalenderToHariLibur {
id?: any;
  kode?: any;
  kdHariLibur?: any;
  kdTanggal?: any;
  kodeExternal?: string;
  namaExternal?: string;
  namaHariLibur?: string;
  reportDisplay?: string;
  statusEnabled?: true;
  tanggal?: any;
  version?: 0;
  noRec?: string;
  label?: any;
  value?: any;


}